Move addToCart &  updateCartQuantity
