+++ amazon.js
@@ -1,5 +1,5 @@
+import {cart,addToCart,updateCartQuantity} from "../data/cart.js";
 import { products } from "../data/products.js";
-import {cart} from "../data/cart.js";
 let productsHTML = '';
 products.forEach((product) => {
   productsHTML += `
@@ -54,34 +54,7 @@ document.querySelectorAll('.js-add-to-cart').forEach((button)=>{
     let addedMessageTimeoutId;
     button.addEventListener('click',()=>{
       const {productId} = button.dataset;
-      let matchingItem;
-      cart.forEach((item)=>{
-        if(productId === item.productId){
-          matchingItem = item;
-        }
-      });
-      const quantitySelector = document.querySelector(`.js-quantity-selector-${productId}`);
-      const quantity = Number(quantitySelector.value);
-      if(matchingItem){
-        matchingItem.quantity += quantity;
-      }else{
-        cart.push({
-          productId,
-          quantity,
-        })
-      }
-      let cartQuantity = 0;
-      cart.forEach((item)=>{
-        cartQuantity += item.quantity;
-      });
-      document.querySelector('.js-cart-quantity').innerHTML = cartQuantity;
-      const addedMessage = document.querySelector(`.js-added-to-cart-${productId}`);
-      addedMessage.classList.add('added-to-cart-visible');
-      if(addedMessageTimeoutId){clearTimeout(addedMessageTimeoutId);
-      }
-      const timeoutId = setTimeout(()=>{
-        addedMessage.classList.remove('added-to-cart-visible')
-      },1500);
-      addedMessageTimeoutId = timeoutId;
+      addToCart(productId);
+      updateCartQuantity();      
     });
   });
